2 new features: 1) support for 'containment-loose' element types.
i.e.,element types that are not targeted by a containment reference of an element which is contained
in the main meta-model will now also recieve CREATE cpeos .
These cpeos are different in that they do not have a connected context node inside.

2.) support for meta-model cross referenced types when using several meta-models (like ow2 using also ecore).

i.e., if types of an additonal meta-model are referenced by types of the main meta-model then these cross
referenced types will also be considered as possible parents/neighbours and imply  additional cpeos.
E.g. now it is possible to generate CREATE_EObject_IN_Ontology. This is a result from the element
Ontology having  a containment reference accross to Ecore's EObject.