update widgets on settings change, some refactoring; details below
DifferenceBuilderWidget and InputModelsWidget are remove from the
*.difference.lifting.ui plugin and the implementation in
*.difference.technical.ui is used instead, as the the former plugin
already depends on the latter one. Imports are updated.

All widgets, that are used to compare models, now update their selection
when the underlying settings change.

Additionally, further problems were fixed in MatchingEngineWidget:
- the order of the matchers is now correct (was undefined behavior
before)
- getSelection() returns null instead of throwing IndexOutOfBounds when
nothing is selected
- the matcher configurations are properly shown/hidden when a
configurable matcher is part of the selection even when multiple
matchers are selected

A method is added to the IncrementalTechnicalDifferenceBuilder to
retrieve the individual difference builders to support the changed
TechnicalDifferenceBuilderWidget.

Version numbers are updated.