refactor IArgumentManager to use PatchingSettings
An argument manager should get all initialization settings directly from
the patching settings object.

To support the retrieval of the most suitable argument manager via its
extension point, the method canResolveArguments now has the settings as
an additional argument and should also check whether the settings are
supported. The method getMode is added to return the mode (interactive,
batch) of an argument manager.