restructure PotentialDependencyAnalyzer, performance improvements
Improve performance of subtype index calculation which was checking the
whole resources for EClasses, which can only appear as children of
EPackages.

Fix logic of isLiteral to use only one call to the factory to check.
Before, it was failing early on the first of multiple checks, which also
turned out to be a minor performance problem.